[raxod502/prescient.el#57] Use string candidates
The Ivy API requires that you use candidates with string values only
in order for sorting and filtering to work properly in all cases. To
provide for non-string candidates there is the option to pass cons
cells whose cars are the strings to use for sorting and filtering and
whose cdrs are the actual values you care about. Doing that fixes a
conflict with prescient.el (see linked issue).